+++ tests/unit_tests/util/__init__.py
@@ -38,10 +38,6 @@ class CommandTestCase:
         else:
             self.parser.print_help()
 
-    def assertGql(self, mock_run, argv, expected_body, expected_vars=None):
-        self.execute(argv)
-        mock_run.assert_called_once_with(expected_body, expected_vars)
-
     def assertRest(
         self, mock_send, argv, method, endpoint, *, body=UNDEFINED, params=UNDEFINED
     ):